Added smart trainerdata.txt handling
Now trainerdata.txt checks if the first line of the file is auto. If it is auto then it tries pulling the data from the loaded save file. If it does not see auto as the first line, it will use the trainerdata.txt in the same way as before via user hardcoded trainer and region values